file-manager/requirements.txt
attrs==20.3.0
boto3==1.17.32
botocore==1.20.32
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==1.1.1
pyrsistent==0.17.3
python-dateutil==2.8.1
s3transfer==0.3.6
six==1.15.0
urllib3==1.26.4
Werkzeug==0.16.1
mysql-connector-python==8.0.23
protobuf==3.15.6
pyjwt==2.0.1
